Remove the old unsued BitabsTranslation, rename new one
